Move magnifier code to separate module
