+++ include/magnifier.h
@@ -1,0 +1,23 @@
+/* SPDX-License-Identifier: GPL-2.0-only */
+#ifndef LABWC_MAGNIFIER_H
+#define LABWC_MAGNIFIER_H
+
+#include <stdbool.h>
+#include <wayland-server-core.h>
+
+struct server;
+struct output;
+
+enum magnify_dir {
+	MAGNIFY_INCREASE,
+	MAGNIFY_DECREASE
+};
+
+void magnify_toggle(struct server *server);
+void magnify_set_scale(struct server *server, enum magnify_dir dir);
+bool output_wants_magnification(struct output *output);
+void magnify(struct output *output, struct wlr_buffer *output_buffer,
+	struct wlr_box *damage);
+bool is_magnify_on(void);
+
+#endif /* LABWC_MAGNIFIER_H */
