+++ src/meson.build
@@ -9,6 +9,7 @@ labwc_sources = files(
   'idle.c',
   'interactive.c',
   'layers.c',
+  'magnifier.c',
   'main.c',
   'node.c',
   'osd.c',
